Rename ProteinSignature -> ProteinSketch and move to sketch.rs
